requirements.txt
streamlit
langdetect
deep-translator
textblob
PyPDF2
python-docx
fpdf2
